Remove unnecessary code from standalone import/export builds
