+++ packages/react-querybuilder/src/utils/parseJsonLogic/parseJsonLogic.ts
@@ -1,5 +1,7 @@
 import type { ReservedOperations } from 'json-logic-js';
-import { filterFieldsByComparator, getValueSourcesUtil, uniqByName } from '../../internal';
+import { filterFieldsByComparator } from '../../internal/filterFieldsByComparator';
+import { getValueSourcesUtil } from '../../internal/getValueSourcesUtil';
+import { uniqByName } from '../../internal/uniq';
 import type {
   DefaultOperatorName,
   DefaultRuleGroupType,
