+++ packages/react-querybuilder/src/utils/toOptions.tsx
@@ -1,7 +1,5 @@
-import type { Field, NameLabelPair, OptionGroup } from '../types/index.noReact';
-
-export const isOptionGroupArray = (arr: Field['values']): arr is OptionGroup[] =>
-  Array.isArray(arr) && arr.length > 0 && 'options' in arr[0];
+import type { NameLabelPair, OptionGroup } from '../types/index.noReact';
+import { isOptionGroupArray } from './optGroupUtils';
 
 export const toOptions = (arr?: NameLabelPair[] | OptionGroup[]) =>
   isOptionGroupArray(arr)
@@ -21,10 +19,3 @@ export const toOptions = (arr?: NameLabelPair[] | OptionGroup[]) =>
         </option>
       ))
     : null;
-
-export const getFirstOption = (arr?: NameLabelPair[] | OptionGroup[]) =>
-  !Array.isArray(arr) || arr.length === 0
-    ? null
-    : isOptionGroupArray(arr)
-    ? arr[0].options[0].name
-    : arr[0].name;
